Move FeedbackRegistryEntry to feedback-components
