+++ src/platform/packages/shared/feedback-components/src/components/body/feedback_body.tsx
@@ -11,11 +11,11 @@ import React from 'react';
 import { EuiFlexGroup, EuiFlexItem, EuiForm } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 
-import type { FeedbackRegistryEntry } from '@kbn/feedback-registry';
 import { FeedbackTextArea } from './feedback_text_area';
 import { EmailSection } from './email';
 import { PrivacyAndSessionDisclaimer } from './privacy_and_session_disclaimer';
 import { CsatButtons } from './csat_buttons';
+import type { FeedbackRegistryEntry } from '../../types';
 
 export interface FeedbackBodyProps {
   selectedCsatOptionId: string;
