+++ src/platform/packages/shared/feedback-components/src/components/feedback_container.tsx
@@ -11,8 +11,7 @@ import React, { useState } from 'react';
 import { EuiFlexGroup, useEuiTheme } from '@elastic/eui';
 import { css } from '@emotion/react';
 import { i18n } from '@kbn/i18n';
-import type { FeedbackRegistryEntry } from '@kbn/feedback-registry';
-import type { FeedbackFormData } from '../types';
+import type { FeedbackFormData, FeedbackRegistryEntry } from '../types';
 import { FeedbackHeader } from './header';
 import { FeedbackBody } from './body/feedback_body';
 import { FeedbackFooter } from './footer/feedback_footer';
