+++ src/platform/packages/shared/feedback-components/src/types.ts
@@ -7,6 +7,45 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
+/**
+ * Definition of a feedback question entry in the feedback registry.
+ */
+export interface FeedbackRegistryEntry {
+  /**
+   * Unique identifier for the feedback entry.
+   */
+  id: string;
+  /**
+   * Sort order for displaying the feedback entry. The lower the number, the higher it appears in the UI.
+   */
+  order: number;
+  /**
+   * The question text to be submitted with telemetry.
+   */
+  question: string;
+  /**
+   * The question text which appears in the UI.
+   */
+  label?: {
+    i18nId: string;
+    defaultMessage: string;
+  };
+  /**
+   * Optional placeholder to show in the UI.
+   */
+  placeholder?: {
+    i18nId: string;
+    defaultMessage: string;
+  };
+  /**
+   * Optional aria-label.
+   */
+  ariaLabel?: {
+    i18nId: string;
+    defaultMessage: string;
+  };
+}
+
 export interface FeedbackQuestion {
   id: string;
   question: string;
