+++ src/platform/packages/shared/feedback-components/tsconfig.json
@@ -11,7 +11,6 @@
     "@kbn/i18n-react",
     "@kbn/test-jest-helpers",
     "@kbn/i18n",
-    "@kbn/feedback-registry",
   ],
   "exclude": [
     "target/**/*",
