+++ x-pack/platform/packages/private/feedback-registry/tsconfig.json
@@ -13,5 +13,7 @@
   "exclude": [
     "target/**/*"
   ],
-  "kbn_references": []
+  "kbn_references": [
+    "@kbn/feedback-components"
+  ]
 }
